# modules/imex/imex-trigger/imex-rest-trigger/imex-rest-trigger-impl/rest-openapi.yaml
info:
    description:
        "This file describes rest api for using IMEX with Rest endpoints. A Java client JAR is available for use with
        the group ID 'com.liferay', artifact ID 'com.liferay.imex.rest.trigger.api.client', and version '1.0.0'."
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
    title: "Imex Rest API"
    version: v1.0
openapi: 3.0.1
paths:
...
  "/reports/files/{fileName}":
    get:
      description: "Download report file identified by fileName"
      parameters:
        - in: path
          name: fileName
          schema:
            type: string
          required: true
          description: File name of the file to get
      responses:
        200:
          description: "File identified by name"
          content:
            application/text:
              schema:
                $ref: "#/components/schemas/ReportFiles"
      tags: ["Reports"]
  "/reports/{identifier}":
    get:
      description: "Gets the list of execution reports as file"
      parameters:
        - in: path
          name: identifier
          schema:
            type: string
          required: true
          description: Identifier of process to get
      responses:
        200:
          description: "Reports files list"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
            application/xml:
              schema:
                $ref: "#/components/schemas/Report"
      tags: ["Reports"]
components:
  schemas:
...
    ReportFiles:
      description: "This is a representation for report logs files"
      properties:
        name:
          description: "Name"
          type: string
        size:
          description: "Size in octets"
          type: string
        lastModifiedDate:
          description: "Date of last modification"
          type: string
          format: date-time
        creationDate:
          description: "Date of creation"
          type: string
          format: date-time 
    Report:
      description: "This is a representation for report"
      properties:
        identifier:
          description: "identifier"
          type: string
        content:
          description: "Report content"
          type: string